+++ beecrypt/rsakp.c
@@ -33,13 +33,13 @@
  * \{
  */
 
-int rsakpMake(rsakp* kp, randomGeneratorContext* rgc, int nsize)
+int rsakpMake(rsakp* kp, randomGeneratorContext* rgc, size_t nsize)
 {
 	/* 
 	 * Generates an RSA Keypair for use with the Chinese Remainder Theorem
 	 */
 
-	register size_t pqsize = ((uint32_t)(nsize+1)) >> 1;
+	register size_t pqsize = (nsize+1) >> 1;
 	register mpw* temp = (mpw*) malloc((16*pqsize+6) * sizeof(*temp));
 	register int newn = 1;
